refactor: 대학교 조회 서비스를 StudentService에서 UniversityService로 별도 분리 (#159)
